Feature - format swap prices (#77)
New price formatting needs to be added to support the new FX Swap price format. This needs to allow an additional decimal pip to be supplied and, optionally, the price to be formatted converted to pips. The changes to support this solution is as follows:

Add new options for specifying number of decimalPips and flag for formatAsPips.

Extend existing decipips logic to handle different decimalPips values.

When formatting as pips:
Round the price to Decimals + DecimalPips decimal places.
Remove the final DecimalPips digits as the DeciPips part.
Convert the remaining price to Pips as price * (10^Decimals)